pyproject.toml
[project]
name = "hardware_collection"
version = "0.1.0"
authors = [
    {name = "Xinkai Jiang", email = "xinkai.jiang@kit.edu"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pytest>=8.3.5",
    "pyzmq>=26.4.0",
]
